Revert "Add sberify"

+++ common/util.go
@@ -1,6 +1,6 @@
 package common
 
-func Contains(slice []int, value int) bool {
+func Contains(slice []int64, value int64) bool {
 	for _, v := range slice {
 		if v == value {
 			return true
